requirements.txt
# Portal Web Corretores - GrupoBand
# Dependências do projeto

Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
